move indexing wrappers to module
