+++ engine/baml-vm/src/lib.rs
@@ -17,5 +17,6 @@ pub(crate) mod vm;
 
 pub use bytecode::{BinOp, Bytecode, CmpOp, Instruction, UnaryOp};
 pub use vm::{
-    BamlVmProgram, Class, Frame, Function, FunctionKind, Instance, Object, ObjectIndex, ObjectPool, StackIndex, EvalStack, Value, Vm, VmExecState,
+    BamlVmProgram, Class, EvalStack, Frame, Function, FunctionKind, Instance, Object, ObjectIndex,
+    ObjectPool, StackIndex, Value, Vm, VmExecState,
 };
